+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2025-10-03
+
+Better default policy for when JSON logs are enabled: machines get JSON, humans get readable logs.
+Enable LogStruct for production servers or when running tests on CI.
+Keep dev-friendly logging on local machines or when running interactive commands on production servers.
+
 ## [0.1.1] - 2025-09-29
 
 Added dotenv-rails integration. Many other fixes and improvements.